Add discrete actions support for GAIL + fix deprecations

+++ stable_baselines/common/distributions.py
@@ -394,7 +394,7 @@ class DiagGaussianProbabilityDistribution(ProbabilityDistribution):
 
     def neglogp(self, x):
         return 0.5 * tf.reduce_sum(tf.square((x - self.mean) / self.std), axis=-1) \
-               + 0.5 * np.log(2.0 * np.pi) * tf.to_float(tf.shape(x)[-1]) \
+               + 0.5 * np.log(2.0 * np.pi) * tf.cast(tf.shape(x)[-1], tf.float32) \
                + tf.reduce_sum(self.logstd, axis=-1)
 
     def kl(self, other):
@@ -438,7 +438,8 @@ class BernoulliProbabilityDistribution(ProbabilityDistribution):
         return tf.round(self.probabilities)
 
     def neglogp(self, x):
-        return tf.reduce_sum(tf.nn.sigmoid_cross_entropy_with_logits(logits=self.logits, labels=tf.to_float(x)),
+        return tf.reduce_sum(tf.nn.sigmoid_cross_entropy_with_logits(logits=self.logits,
+                                                                     labels=tf.cast(x, tf.float32)),
                              axis=-1)
 
     def kl(self, other):
@@ -453,7 +454,7 @@ class BernoulliProbabilityDistribution(ProbabilityDistribution):
 
     def sample(self):
         samples_from_uniform = tf.random_uniform(tf.shape(self.probabilities))
-        return tf.to_float(math_ops.less(samples_from_uniform, self.probabilities))
+        return tf.cast(math_ops.less(samples_from_uniform, self.probabilities), tf.float32)
 
     @classmethod
     def fromflat(cls, flat):
